Add Infinite mode after Amulet level
--HG--
branch : khaotic

+++ src/com/kbuffer/khaoticpixeldungeon/actors/mobs/DM300.java
@@ -133,9 +133,14 @@ public class DM300 extends Mob {
 		
 		super.die( cause );
 		
-		GameScene.bossSlain();
-		Dungeon.level.drop( new SkeletonKey( Dungeon.depth  ), pos ).sprite.drop();
-		
+		// KPD - Infinite mode
+		// Don't do this stuff later on
+		if( Dungeon.depth < 26 ) {
+			GameScene.bossSlain();
+			Dungeon.level.drop(new SkeletonKey(Dungeon.depth), pos).sprite.drop();
+		}
+		// end KPD
+
 		Badges.validateBossSlain();
 
 		LloydsBeacon beacon = Dungeon.hero.belongings.getItem(LloydsBeacon.class);
