+++ src/com/kbuffer/khaoticpixeldungeon/actors/mobs/Goo.java
@@ -199,9 +199,14 @@ public class Goo extends Mob {
 		
 		Dungeon.level.unseal();
 		
-		GameScene.bossSlain();
-		Dungeon.level.drop( new SkeletonKey( Dungeon.depth ), pos ).sprite.drop();
-		
+		// KPD - Infinite mode
+		// Don't do this stuff later on
+		if( Dungeon.depth < 26 ) {
+			GameScene.bossSlain();
+			Dungeon.level.drop(new SkeletonKey(Dungeon.depth), pos).sprite.drop();
+		}
+		// end KPD
+
 		Badges.validateBossSlain();
 		
 		yell( "glurp... glurp..." );
